+++ src/main/java/com/example/apibimo/repository/CurriculoRepository.java
@@ -2,6 +2,6 @@ package com.example.apibimo.repository;
 import com.example.apibimo.models.Curriculo;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CurriculoRepository extends JpaRepository<Curriculo, Integer> {
+public interface CurriculoRepository extends JpaRepository<Curriculo, Long> {
 
 }
